:recycle: split loader/saver
