+++ caikit/core/__init__.py
@@ -32,14 +32,16 @@ _warnings.filterwarnings("ignore")
 # Local
 from .data_model import DataObjectBase, dataobject
 from .model_manager import *
-from .modules.base import ModuleBase, module
-from caikit.core.module_backends.module_backend_config import configure as backend_configure
 from .module_backends import *
-from caikit.core.modules.config import ModuleConfig
+from .modules.base import ModuleBase, module
 from .resources.base import ResourceBase, resource
 from .task import TaskBase, task
 from .toolkit import *
 from .workflows.base import WorkflowBase, workflow
+from caikit.core.module_backends.module_backend_config import (
+    configure as backend_configure,
+)
+from caikit.core.modules.config import ModuleConfig
 
 # Configure the global model wrangling functions
 MODEL_MANAGER = ModelManager()
