+++ caikit/core/workflows/base.py
@@ -27,8 +27,8 @@ import alog
 # Local
 from ... import core
 from .. import module as mod
-from caikit.core.modules.decorator import module_type
 from ..toolkit.errors import error_handler
+from caikit.core.modules.decorator import module_type
 
 log = alog.use_channel("WFBASE")
 error = error_handler.get(log)
